+++ src/parser/inline/html.ts
@@ -41,11 +41,6 @@ export const html: HTMLParser = lazy(() => creator(validate('<', union([
         startTight(
         context((() => {
           switch (tag) {
-            case 'bdo':
-            case 'bdi':
-              return {
-                state: { in: { bdx: true } }
-              };
             case 'sup':
             case 'sub':
               return {
@@ -56,10 +51,11 @@ export const html: HTMLParser = lazy(() => creator(validate('<', union([
                 }},
               };
             case 'small':
-            default:
               return {
                 state: { in: { small: true } }
               };
+            default:
+              return {};
           }
         })(),
         some(union([inline]), `</${tag}>`))),
@@ -97,13 +93,6 @@ export const attribute: HTMLParser.TagParser.AttributeParser = union([
 function elem(tag: string, as: (HTMLElement | string)[], bs: (HTMLElement | string)[], cs: (HTMLElement | string)[], context: MarkdownParser.Context): HTMLElement {
   if (!tags.includes(tag)) return invalid('tag', `Invalid HTML tag <${tag}>.`, as, bs, cs);
   switch (tag) {
-    case 'bdo':
-    case 'bdi':
-      switch (true) {
-        case context.state?.in?.bdx:
-          return invalid('nest', `<${tag}> HTML tag can't be used in <bdo>/<bdi> HTML tags.`, as, bs, cs);
-      }
-      break;
     case 'sup':
     case 'sub':
       switch (true) {